+++ lib/websockets/interfaces.ts
@@ -6,7 +6,7 @@ import type { Node } from 'prosemirror-model';
 import type { Server, Socket } from 'socket.io';
 
 import type { UIBlockWithDetails } from 'lib/databases/block';
-import type { FailedImportsError } from 'lib/notion/types';
+import type { FailedImportsError } from 'lib/notion/interfaces';
 import type { ProposalWithUsersLite } from 'lib/proposals/getProposals';
 import type { RewardBlockWithTypedFields } from 'lib/rewards/blocks/interfaces';
 import type { ExtendedVote, VoteTask } from 'lib/votes/interfaces';
