+++ package.json
@@ -138,7 +138,7 @@
     "@mux/mux-video-react": "^0.7.15",
     "@mux/upchunk": "^3.0.0",
     "@noble/ed25519": "^2.0.0",
-    "@notionhq/client": "^2.2.2",
+    "@notionhq/client": "^2.2.15",
     "@octokit/app": "^14.1.0",
     "@octokit/auth-app": "^6.1.1",
     "@octokit/auth-oauth-app": "^8.1.0",
